Fix SMB passwords containing special characters being unusable
When an SMB password contained shell control characters like $! or the %
sign used to separate the username from the password with the
--user=user%password flag, the called smbclient command would crash and
the web interface would tell the user that no shares could be found.

We now specify the password with the --password flag instead of the
--user flag and enclose it in single quotes to allow all characters.
Only the literal single quote has to be escaped to not trip up the shell
executing the command.